Create FizzBuzz
The classic FizzBuzz programming assignment with a SoloLearn Twist! In this assignment, this code takes an input n and outputs the numbers from 1 to n. For each multiple of 3, it prints "Solo" instead of the number. For each multiple of 5, it prints "Learn" instead of the number. For numbers that multiples of both 3 and 5, it outputs "SoloLearn". In addition to that, the code skips even numbers so that the logic only applies to odd numbers in the range.